refactor(layout): layout context를 layout 하위로 이동하고 layout provider 디렉토리 제거
