+++ src/layout/LayoutContext.ts
@@ -3,8 +3,8 @@ import { createContext, Dispatch } from 'react'
 import { noop } from 'lodash-es'
 
 /* Internal dependencies */
-import { LayoutState } from '../layout/redux/LayoutReducer'
-import { LayoutActionTypes } from '../layout/redux/LayoutActions'
+import { LayoutState } from './redux/LayoutReducer'
+import { LayoutActionTypes } from './redux/LayoutActions'
 
 export const LayoutStateContext = createContext<LayoutState | undefined>(undefined)
