--- src/layout/LayoutProvider/index.ts
@@ -1,5 +1,0 @@
-import LayoutProvider from './LayoutProvider'
-
-export {
-  LayoutProvider,
-}
