+++ src/layout/stories/LayoutPlayground/LayoutPlayGround.stories.tsx
@@ -10,7 +10,7 @@ import { getTitle } from '../../../utils/storyUtils'
 import { styled, Typography } from '../../../foundation'
 import { Icon, IconSize } from '../../../components/Icon'
 import { ListItem } from '../../../components/ListItem'
-import { LayoutProvider } from '../../LayoutProvider'
+import LayoutProvider from '../../LayoutProvider'
 import Client from '../../Client/Client'
 import { Main } from '../../Main'
 import GNB from '../../GNB/GNB'
